* commit '63d3d0ba1332a6124ef6376848c11e223e76cfb7':
  Fix destream -Wconversion warnings.

+++ framework/delibs/destream/deThreadStream.c
@@ -48,7 +48,7 @@ static void inStreamCopy (void* arg)
 {
 	deThreadInStream* threadStream = (deThreadInStream*)arg;
 
-	deUint8* buffer = malloc(sizeof(deUint8) * threadStream->bufferSize);
+	deUint8* buffer = malloc(sizeof(deUint8) * (size_t)threadStream->bufferSize);
 
 	for(;;)
 	{
